refactor: update error handling
Apparently the first argument from request module's callback is not returning any error when the request failed. So I added a new checking for body.errors to make sure that all errors are handled apart from the error arg.

One more addtion is the message field in the error object. Since paymongo's api is based on json api standard, the message of the error is stored in the `detail` field.